moving shadow mappings back
